move pkg/storage/hash to pkg/pachhash, and use blake2b
